+++ app/models/project.rb
@@ -3,6 +3,8 @@ class Project < ApplicationRecord
   has_many :notes, dependent: :destroy
   has_many :contacts, dependent: :destroy
 
+  accepts_nested_attributes_for :notes, reject_if: proc { |attributes| attributes[:body].blank? }
+
   validates :urn, presence: true, numericality: {only_integer: true}, length: {is: 6}
   validates :trust_ukprn, presence: true, numericality: {only_integer: true}
   validates :target_completion_date, presence: true